+++ app/src/main/java/io/github/muntashirakon/AppManager/db/dao/FreezeTypeDao.java
@@ -5,16 +5,20 @@ package io.github.muntashirakon.AppManager.db.dao;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.room.Dao;
-import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 
+import java.util.List;
+
 import io.github.muntashirakon.AppManager.db.entity.FreezeType;
-import io.github.muntashirakon.AppManager.utils.FreezeUtils;
 
 @Dao
 public interface FreezeTypeDao {
+    @NonNull
+    @Query("SELECT * FROM freeze_type")
+    List<FreezeType> getAll();
+
     @Nullable
     @Query("SELECT * FROM freeze_type WHERE package_name = :packageName LIMIT 1")
     FreezeType get(String packageName);